+++ cdk-morphlines/cdk-morphlines-core/src/main/java/com/cloudera/cdk/morphline/stdlib/GrokBuilder.java
@@ -43,7 +43,7 @@ import com.typesafe.config.ConfigFactory;
 public final class GrokBuilder implements CommandBuilder {
 
   /*
-   * Uses a shaded version of com.google.code.regexp-0.1.9 to minimize potential dependency issues.
+   * Uses a shaded version of com.google.code.regexp-0.2.3 to minimize potential dependency issues.
    * See https://github.com/tony19/named-regexp
    */