+++ src/site/markdown/release_notes.md
@@ -17,6 +17,7 @@ Version TBD has the following notable changes:
     * Added option to specify boost values to `loadSolr` command.
     * Added several performance enhancements.
     * Upgraded `cdk-morphlines-solr-cell` maven module from tika-1.3 to tika-1.4 to pick up some bug fixes.
+    * Upgraded `cdk-morphlines-core` maven module from com.google.code.regexp-0.1.9 to 0.2.3 to pick up some bug fixes (Internally shaded version).
     * The constructor of AbstractCommand now has an additional parameter that refers to the CommandBuilder. 
     * The ISO8601_TIMEZONE grok pattern now allows the omission of minutes in a timezone offset.
     * Ensured morphline commands can refer to record field names containing arbitrary characters.